requirements.txt
requests
beautifulsoup4
psycopg[binary]
python-dotenv
pandas
jupyter
nbformat
